optimizations and front-end development

+++ app/Http/Requests/ProductFormRequest.php
@@ -46,11 +46,11 @@ class ProductFormRequest extends FormRequest
             ],
             'original_price' => [
                 'required',
-                'integer',
+                'numeric',
             ],
             'selling_price' => [
                 'required',
-                'integer',
+                'numeric',
             ],
             'quantity' => [
                 'required',
